+++ src/main/java/ru/skypro/homework/config/WebSecurityConfig.java
@@ -22,6 +22,7 @@ import javax.sql.DataSource;
 import static org.springframework.security.config.Customizer.withDefaults;
 
 @Configuration
+@EnableWebSecurity
 public class WebSecurityConfig {
 
     private static final String[] AUTH_WHITELIST = {
@@ -65,9 +66,9 @@ public class WebSecurityConfig {
 
 
 
-    @Bean
+   /* @Bean
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
-
+*/
 }
