--- src/main/java/ru/skypro/homework/testSecurityConfig/ExperimentFilter.java
@@ -1,39 +1,0 @@
-package ru.skypro.homework.testSecurityConfig;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.nio.file.AccessDeniedException;
-import java.util.logging.Filter;
-import java.util.logging.LogRecord;
-
-public class ExperimentFilter implements Filter {
-
-    public void doFilter(ServletRequest servletRequest,
-                         ServletResponse servletResponse,
-                         FilterChain filterChain) throws IOException, ServletException {
-        HttpServletRequest request = (HttpServletRequest) servletRequest;
-        HttpServletResponse response = (HttpServletResponse) servletResponse;
-
-        String tenantId = request.getHeader("X-Tenant-Id");
-        boolean hasAccess = isLoggable(tenantId);
-        if (hasAccess) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-        throw new AccessDeniedException("Access denied");
-    }
-
-    public boolean isLoggable(String record) {
-        return false;
-    }
-
-    @Override
-    public boolean isLoggable(LogRecord record) {
-        return false;
-    }
-}
